Refactor SIMD ifdefs
The extconf check if SIMD is available, but then later on
the `simd.h` header has extra logic to decide if it uses
it or not.

e.g. for SSE2 we have:

```
if defined(__amd64__) || defined(__amd64) || defined(__x86_64__) ||
   defined(__x86_64) || defined(_M_X64) || defined(_M_AMD64)
```

Which exclude i686 builds.

Hence if we're not entering that `#if` we can end up with
`JSON_ENABLE_SIMD=1` but neither `HAVE_SIMD_NEON` nor `HAVE_SIMD_SSE2`.

So instead of relying on `JSON_ENABLE_SIMD` to check if any SIMD
implementation is available, this commit introduce `HAVE_SIMD`.