Better handle missing ostruct
In the Ruby test suite, this test class is causing trouble because
ostruct is not available. Having an autoload for JSON::GenericObject but
causing it not to define the constant causes a warning.

See 0dc1cd407e7775610f2bcaef6c1282369867f91c and
caa5d8cdd7483647013af5e3d1701b0ed3ec8785 in ruby.

We can skip defining the test class entirely instead when ostruct is not
available.